add python wrapper

+++ sobel.h
@@ -4,5 +4,5 @@
 
 #ifndef UNTITLED1_SOBEL_H
 #define UNTITLED1_SOBEL_H
-void sobel(int* bins, double* data, double* output, int nbins, int nrows, int* nBinsInRow, int* basebins, int fillValue);
+void sobel(int* bins, double* data, double* output, int nbins, int nrows, int* nBinsInRow, int* basebins, double fillValue);
 #endif //UNTITLED1_SOBEL_H
